Added login page and modified bio pages
The login page asks for an email and password, then allows the user to login with an existing account or create a new account with the entered credentials.

The 3 bio pages have become more unified so that only one html file has to be modified to change the layout of all 3 pages.  This also fixed a bug that caused the data to disappear on reload (only showed when the page was clicked in the menu).